requirements.txt
# Core dependencies for BMAD Portfolio Management System

# Angel One API
smartapi-python==1.3.0
pyotp==2.9.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Market data
yfinance==0.2.33

# Technical indicators
TA-Lib==0.4.28
pandas-ta==0.3.14b0

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# API Framework
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# HTTP requests
requests==2.31.0

# Database
sqlalchemy==2.0.25

# LLM for sentiment
openai==1.6.1
anthropic==0.8.1

# Memory system for agents
memorisdk>=2.0.1
loguru>=0.6.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1
